Make download directory a variable
defaults/main.yml:
    * Add variable `gitlab_ci_multi_runner_dl_dir`, default to /tmp

tasks/main.yml:
    * Use gitlab_ci_multi_runner_dl_dir variable

At times, the machine that we run the Ansible playbook against does
not have /tmp mounted exec. So we need to have the install script
stored at other place. Alternatively, we can use something like
"bash gitlab-runner.sh".

It is not the VM that we want to Ansible but the machine that we have
to.

Signed-off-by: Tuan T. Pham <tuan@vt.edu>